feat(wiring, AbstractComponent): Simplify wiring and standardize API to be more like Component
The wiring API was not in line with the Component API. This change standardizes the wireIn
function to "wire", and creates an Atomic.Wire function which behaves similar syntactically
to Atomic.Component.

BREAKING CHANGE: In order to support the new API, all wirings need to be udpated from originally
just calling a define() method to including the Atomic/Wiring component and returning the call
to Wiring().

Previously:
(function(define) {
  define('wirings/foo', [], function() {
    return function(config) { /*...*/ };
  });
}(/*amd signature*/));

New syntax:
(function(define) {
  define('wirings/bottoming', ['Atomic/Wiring'], function(Wiring) {
    return Wiring(function(config) { /*...*/ });
  });
}(/*amd signature*/));