feat(PublicAPI): Adds powerful debugging features to Components
This adds Atomic.describe(component, output) to the Atomic Public API.
Previously, debugging was powerful in console.log, but there was no
simple API for getting a component's description and attributes. This
exposes all the displayable utilities via a single interface. If given
a string, it will download the component and display output. If given
a component object, it will use that component for description.

A future revision may wish to also enable typeof === 'function', so
that uninstantiated components can also be described.

A promise is returned which when thenned will expose a JSON representation
of the component's description, ideal for Dust/Handlebars/Template
language of choice for documentation pages.

BREAKING CHANGE: In order to ensure documentation works as expected,
the following methods have had restrictions added:
* trigger: the event name must be specified in the component.events collection
* assign: the element property must be specified in component.elements
* resolve: the dependency must be specified in component.depends
* state: the state being set must be specified in component.states

This affects all components utilizing the trigger/assign/resolve/state methods