refactor(grunt): Cleans up grunt file to add semver support
Moves everything around a lot. Basically taking advantage of improvements
in grunt plugins to better organize the code base. This should make it
easier to find and test various pieces.

Each module behaves independently now with a self executing function that
takes an Atomic variable. Dependencies can now be injected by passing in
a generic / custom Atomic object for the various pieces.

Semver was added as part of the builds as part of the planned feature
development for 0.10.0.